Factor view all click handler into separate class
